requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10

aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
assemblyai==0.40.2
asteroid-filterbanks==0.4.0
async-timeout==5.0.1
attrs==24.2.0
audeer==2.2.0
audiofile==1.5.0
audioread==3.0.1
audmath==1.4.1
certifi==2024.8.30
cffi==1.17.1
facenet-pytorch==2.6.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
datasets==3.1.0
decorator==4.4.2
dill==0.3.8
distro==1.9.0
docopt==0.6.2
einops==0.8.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.6
filelock==3.16.1
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.9.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
HyperPyYAML==1.2.2
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.11.9
llvmlite==0.43.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
openai==1.57.2
opencv-python==4.10.0.84
openpyxl==3.1.5
optuna==4.1.0
packaging==24.2
pandas==2.2.3
pillow==10.2.0
platformdirs==4.3.6
pooch==1.8.2
primePy==1.3


pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1

python-dateutil==2.9.0.post0
python-dotenv==1.0.1


safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.5.0.post1
speechbrain==1.0.2
SQLAlchemy==2.0.36
starlette==0.41.3
sympy==1.13.1
python-multipart
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.2.0
torch-audiomentations==0.11.1
torch_pitch_shift==1.2.5
torchaudio==2.2.0
torchmetrics==1.6.1
tqdm==4.67.1
transformers==4.47.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
websockets==14.1
xxhash==3.5.0
yarl==1.18.3
mediapipe==0.10.20
